Add more scripting documentation

+++ src/scripting/src/interop/modules/imppg.cpp
@@ -117,7 +117,7 @@ const luaL_Reg functions[] = {
     {"progress", [](lua_State* lua) -> int {
         CheckNumArgs(lua, "progress", 1);
         const int percentage = GetNumber(lua, 1);
-        scripting::g_State->SendMessage(scripting::contents::Progress{percentage});
+        scripting::g_State->SendMessage(scripting::contents::Progress{static_cast<double>(percentage)});
         return 0;
     }},
